pyproject.toml
[project]
name = "vl-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["openai>=1.83.0", "python-dotenv>=1.1.0"]

[build-system]
requires = ["setuptools>=42", "wheel", "uv>=0.1.0"]
build-backend = "setuptools.build_meta"
